+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 
 - Added testing infrastructure (pytest, GitHub Action).
 
-- Added tests for series reading.
+- Added tests for party, series reading.
 
 - Added ruff as a development dependency.